+++ pages/about.jsx
@@ -1,9 +1,9 @@
 // aboutページ
 
-import Meta from '@/src/components/elements/meta/Meta'
-import Container from '@/src/components/layouts/container/Container'
-import Profile from '@/src/components/profile/Profile'
-import Hero from '@/src/components/elements/hero/Hero'
+import Meta from '@/src/components/Elements/meta/Meta'
+import Container from '@/src/components/Layouts/Container/Container'
+import Profile from '@/src/components/Profile/Profile'
+import Hero from '@/src/components/Elements/hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/about.jpg'
