+++ pages/articles/[slug].jsx
@@ -1,23 +1,23 @@
 // 記事ページ
 
 import { getPostBySlug, getAllSlugs, getAllCategories } from '@/libs/api'
-import Container from '@/src/components/layouts/container/Container'
+import Container from '@/src/components/Layouts/Container/Container'
 import PostHeader from '@/pages/articles/components/PostHeader/PostHeader'
 import Image from 'next/image'
 import {
   TwoColum,
   TwoColumMain,
   TwoColumSidebar,
-} from '@/src/components/layouts/two-colum/TwoColum'
-import ConvertBody from '@/src/components/convert/ConvertBody'
+} from '@/src/components/Layouts/TwoColum/TwoColum'
+import ConvertBody from '@/src/components/Convert/ConvertBody'
 import PostCategories from '@/pages/articles/components/PostCategories/PostCategories'
 import { extractText } from '@/libs/extract-text'
-import Meta from '@/src/components/elements/meta/Meta'
+import Meta from '@/src/components/Elements/meta/Meta'
 import { prevNextPost } from '@/libs/prev-next-post'
-import Pagination from '@/src/components/elements/pagination/Pagination'
+import Pagination from '@/src/components/Elements/pagination/Pagination'
 import PostBody from '@/pages/articles/components/PostBody/PostBody'
 import { renderToc } from '@/libs/render-toc'
-import TableOfContents from '@/src/components/elements/table-of-contents/TableOfContents'
+import TableOfContents from '@/src/components/Elements/table-of-contents/TableOfContents'
 import 'highlight.js/styles/night-owl.css'
 
 export default function Post({
