+++ pages/index.jsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from 'react';
 // import { Link, Element, animateScroll as scroll } from 'react-scroll';
 import { getAllPosts } from '@/libs/api'
-import Meta from '@/src/components/elements/meta/Meta'
-import Container from '@/src/components/layouts/container/Container'
-import Hero from '@/src/components/elements/hero/Hero'
+import Meta from '@/src/components/Elements/meta/Meta'
+import Container from '@/src/components/Layouts/Container/Container'
+import Hero from '@/src/components/Elements/hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/index.jpg'
-import TopAbout from '@/src/components/top-about/top-about'
+import TopAbout from '@/src/components/TopAbout/TopAbout'
 import Posts from '@/pages/articles/components/Posts/Posts'
 
 export default function Home({ posts }) {
