+++ pages/works.jsx
@@ -1,8 +1,8 @@
 // worksページ
 
-import Meta from '@/src/components/elements/meta/Meta'
-import Container from '@/src/components/layouts/container/Container'
-import Hero from '@/src/components/elements/hero/Hero'
+import Meta from '@/src/components/Elements/meta/Meta'
+import Container from '@/src/components/Layouts/Container/Container'
+import Hero from '@/src/components/Elements/hero/Hero'
 import eyecatch from 'images/works.jpg'
 import Image from 'next/image'
