+++ src/components/layouts/layout.jsx
@@ -1,7 +1,7 @@
 // ヘッダーとフッターをLayoutコンポーネントで管理する
 
-import Header from './header/Header'
-import Footer from './footer/Footer'
+import Header from './Header/Header'
+import Footer from './Footer/Footer'
 
 export default function Layout({ children }) {
   return (
